extract address bus
