+++ gb-bus/src/generic.rs
@@ -1,0 +1,5 @@
+mod chardev;
+mod random;
+
+pub use chardev::CharDevice;
+pub use random::RandomDevice;
