+++ gb-bus/src/lib.rs
@@ -1,6 +1,11 @@
-use crate::address_bus::Error;
+pub mod address_bus;
+pub mod file_operation;
+pub mod generic;
 
-trait AddressBus {
+pub use address_bus::{Address, AddressBus, Area, Error};
+pub use file_operation::FileOperation;
+
+trait Bus {
     /// read `N` into the bus where `N: u8 | u16`
     fn read<N>(address: u16) -> Result<N, Error>;
     /// write `N` into the bus where `N: u8 | u16`
