--- gb-cpu/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod address_bus;
-pub mod r#trait;
-
-pub use address_bus::{AddressBus, Area, FileOperation};
